% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorddiag-package.R
\docType{package}
\name{chorddiag-package}
\alias{chorddiag-package}
\title{chorddiag: A package for creating interactive chord diagrams}
\description{
Create interactive chord diagrams using the JavaScript visualization library
D3 (\url{http://d3js.org}). The chord diagram layout is based on
\url{http://bl.ocks.org/mbostock/4062006}. Chord diagrams show directed
relationships among a group of entities.
}
