% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chorddiag.R
\name{chorddiag}
\alias{chorddiag}
\title{Create a D3 Chord Diagram}
\source{
\url{http://bl.ocks.org/mbostock/4062006}
}
\usage{
chorddiag(data, width = NULL, height = NULL, palette = "Set2",
  groupColors = NULL, groupPadding = 0.05, groupnamePadding = NULL,
  groupnameFontsize = 18, showTicks = TRUE, tickInterval = NULL,
  ticklabelFontsize = 10)
}
\arguments{
\item{data}{A square matrix containing the data. Group names to be used as
labels must be set via \code{colnames}}

\item{width}{Width for the chord diagram's frame area in pixels (if NULL then
eidth is automatically determined based on context).}

\item{height}{Height for the chord diagram's frame area in pixels (if NULL
then height is automatically determined based on context).}

\item{palette}{A character string. The name of the colorbrewer palette to be
used.}

\item{groupColors}{A vector of colors to be used for the groups. If provided
overwrites the palette.}

\item{groupPadding}{A numeric.}

\item{groupnamePadding}{A numeric.}

\item{groupnameFontsize}{Numeric font size in pixels for the group labels.}

\item{showTicks}{A logical scalar.}

\item{tickInterval}{A numeric value.}

\item{ticklabelFontsize}{Numeric font size in pixels for the tick labels.}
}
\description{
Create an interactive chord diagram using the JavaScript visualization
library D3 (\url{http://d3js.org}). More specifically, the chord diagram
layout is based on \url{http://bl.ocks.org/mbostock/4062006}
}
\examples{
m <- matrix(c(11975,  5871, 8916, 2868,
               1951, 10048, 2060, 6171,
               8010, 16145, 8090, 8045,
               1013,   990,  940, 6907),
               byrow = TRUE,
               nrow = 4, ncol = 4)
groupnames <- c("black", "blonde", "brown", "red")
row.names(m) <- groupnames
colnames(m) <- groupnames
chorddiag(m)
}

