% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chorddiag.R
\name{chorddiagOutput}
\alias{chorddiagOutput}
\title{Widget output function for use in Shiny}
\usage{
chorddiagOutput(outputId, width = "90\%", height = "350px")
}
\arguments{
\item{outputId}{Shiny output variable to read from.}

\item{width}{A valid CSS unit for the width of the chord diagram output.}

\item{height}{A valid CSS unit for the height of the chord diagram output.}
}
\description{
Widget output function for use in Shiny
}
