% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chorddiag.R
\name{chorddiag}
\alias{chorddiag}
\title{Create a D3 Chord Diagram}
\source{
\url{http://bl.ocks.org/mbostock/4062006}
}
\usage{
chorddiag(data, type = "directional", width = NULL, height = NULL,
  margin = 100, palette = "Set2", groupNames = NULL, groupColors = NULL,
  groupThickness = 0.1, groupPadding = 0.05, groupnamePadding = NULL,
  groupnameFontsize = 18, groupedgeColor = NULL,
  chordedgeColor = "#808080", categoryNames = NULL,
  categorynameFontsize = 20, showTicks = TRUE, tickInterval = NULL,
  ticklabelFontsize = 10, fadeLevel = 0.1, showTooltips = TRUE,
  tooltipUnit = NULL, tooltipGroupConnector = " &#x25B6; ",
  precision = NULL)
}
\arguments{
\item{data}{A matrix containing the data. Must be square for the
"directional" type. Column names of the matrix (if existing) will be used
as group labels unless the \code{groupNames} argument is explicitely set.
For the "bipartite" type, the column names label the groups on the left
side of the chord diagram whereas the row names label the groups on the
right side.}

\item{type}{A character string for the type of chord diagram. Either
"directional" (default) or "bipartite". Chord diagrams can be helpful for
visualising symmetric relations between two categories of groups, too.}

\item{width}{Width for the chord diagram's frame area in pixels (if NULL then
eidth is automatically determined based on context).}

\item{height}{Height for the chord diagram's frame area in pixels (if NULL
then height is automatically determined based on context).}

\item{margin}{Numeric margin in pixels between the outer diagram radius and
the edge of the display.}

\item{palette}{A character string. The name of the colorbrewer palette to be
used.}

\item{groupNames}{A vector of character strings to be used for group
labeling.}

\item{groupColors}{A vector of colors to be used for the groups. Providing
\code{groupColors} overrides any \code{palette} given.}

\item{groupThickness}{Numeric thickness for the groups as a fraction of the
total diagram radius.}

\item{groupPadding}{Numeric padding between groups.}

\item{groupnamePadding}{Numeric padding between diagram and group labels. Use
this argument if group labels overlag with tick labels.}

\item{groupnameFontsize}{Numeric font size in pixels for the group labels.}

\item{groupedgeColor}{Color for the group edges. If NULL group colors will be
used.}

\item{chordedgeColor}{Color for the chord edges.}

\item{categoryNames}{A length-2 vector of character strings to be used for
category labels (left and right side of a bipartite chord diagram).}

\item{categorynameFontsize}{Numeric font size in pixels for the category
labels in a bipartite diagram.}

\item{showTicks}{A logical scalar.}

\item{tickInterval}{A numeric value.}

\item{ticklabelFontsize}{Numeric font size in pixels for the tick labels.}

\item{fadeLevel}{Numeric chord fade level (opacity value between 0 and 1,
defaults to 0.1).}

\item{showTooltips}{A logical scalar.}

\item{tooltipUnit}{A character string for the units to be used in tooltips.}

\item{tooltipGroupConnector}{A character string to be used in tooltips:
"<source group> <tooltipGroupConnector> <target group>". Defaults to a
triangle pointing from source to target.}

\item{precision}{Integer number of significant digits to be used for tooltip
  display.}
}
\description{
Create an interactive chord diagram using the JavaScript visualization
library D3 (\url{http://d3js.org}). More specifically, the chord diagram
layout is based on \url{http://bl.ocks.org/mbostock/4062006}. Chord diagrams
show directed relationships among a group of entities.
}
\examples{
m <- matrix(c(11975,  5871, 8916, 2868,
               1951, 10048, 2060, 6171,
               8010, 16145, 8090, 8045,
               1013,   990,  940, 6907),
               byrow = TRUE,
               nrow = 4, ncol = 4)
groupnames <- c("black", "blonde", "brown", "red")
row.names(m) <- groupnames
colnames(m) <- groupnames
chorddiag(m)
}

