% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chorddiag.R
\name{chorddiagOutput}
\alias{chorddiagOutput}
\title{Widget output function for use in Shiny}
\usage{
chorddiagOutput(outputId, width = "90\%", height = "350px")
}
\description{
Widget output function for use in Shiny
}

