% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chorddiag.R
\name{renderChorddiag}
\alias{renderChorddiag}
\title{Widget render function for use in Shiny}
\usage{
renderChorddiag(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

